Fix Makefile bug that was setting PLATFORM_DESC incorrectly.
It had two values in it; 'include' was added, because we were
grepping for that string but not removing it afterwards.

Now we use awk to pull the value more efficiently.

This may have been causing unnecessary rebuilds if the 'include'
directory had its time updated.

With this change, the MACHINE_FILE setting is also ignored,
and defaults to the machine name plus the .md extension.