Speed up device update in some error cases.
When one of these 3 things happens:
- a kick request is delayed
- a kickout lock is held (similar)
- trough count abruptly changes at end of update

then we restart update, but there is no need to wait for
the switch settle delay (since no switch activity necessarily
happened).  A small delay of 50ms is inserted just to keep
the scheduler happy.  This gets rid of some unnecessary delay
when these things happen for a brief time.